Change 'package' in manifest to 'it.sms1920.spqs.ufit.view'
The attribute 'package' of the manifest defines where the class R will
be instantiated. Since in MVP resources must be accessed only from the
components of the view, the previous value of 'package' made necessary
to import R from the package 'it.sms1920.spqs.ufit'. Remove the
useless imports and remove any access to resources from the presenter
and model components.